Fixed penguins reserving multiple entities
This bug was causing penguins to reserve stockpiles and
entities multiple times in a single turn even though the
penguins would only use one of the reserved items.
Because of this bug, whole sections of stockpiles were
unusable and many fish would be permanently "reserved"
so noone would touch them.

The problem should be fixed now.